Rename "openwisk" to "openwhisk" module
