+++ deepaas/cmd/wsk.py
@@ -23,7 +23,7 @@ from oslo_log import log as logging
 import deepaas
 from deepaas.cmd import _shutdown
 from deepaas import config
-from deepaas.openwisk import proxy
+from deepaas.openwhisk import proxy
 
 cli_opts = [
     cfg.StrOpt('listen-ip',
